[!!!][TASK] Restructure cache handling
The storing of the cached PDF files is consolidates in the
CacheManager utility class. The filenames now match the
cache identifiers to allow cleanup of expired entries.

Since we want to use an expiration time and tags the
SimpleFileBackend is replaced with the DatabaseBackend.

Since we do not store any complex data the StringFrontend is
used instead of the VariableFrontend.

The folder for the cached PDF files is now configured in the
Extension configuration to make it available to future cleanup
tasks.

To improve security the cache identifiers are now generated using
the hmac() functionality. Additionally the cache directory is
automatically protected with an .htaccess file.