[FEATURE] Improved cache garbage collection
A custom database backend is used for the PDF cache that
allows the configuration of a maximum number of cache entries.
If this limit is exceeded, old entries are deleted when a new
entry is added.

This custom backend will also execute a garbage collection for
the cached PDF files. All files that do not have a matching
valid cache entry are deleted. To profit from this new
cleanup function the caching framework garbage collection task
needs to be configured to clear the cache for the new Backend.